requirements.txt
Flask
Flask-WTF
Flask-Login
Flask-SQLAlchemy
SQLAlchemy-Datatables
Bootstrap-Flask
ibm-db
ibm-db-sa
